Moved session storage utility to sessionStorage.ts
